refactor: move redis feature to src directory
